back/requirements.txt
Django==1.9.7
django-channels==0.6.0
django-cors-headers==1.1.0
django-extensions==1.6.7
django-filters==0.2.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
